Test for individual candidate existance
Previously, Template::destination would enumerate
files order to find the next available name for a
backup.

However, filesystems vary in terms of their notion
of nominal equality. For example, filesystems may
be:
- ascii case (in)sensitive
- unicode case (in)sensitive
- unicode normalization (in)sensitive

Given this, it's probably best to instead check
the existance of individual potential candidates,
and thus defer to the filesystems own notion of
string equality.

This patch does just that. Unfortunately, bak will
now fill holes in sequences of backups, which I
don't think is ideal, but so it goes.