Allow the user to pass a format string, not only extension
This new `Extension` struct allows users to have control over how the
duplicate count number appears.

Given MFEK works on UFO fonts, this commit is vital for us:

Example.ufo ⇒ Example.bak.0.ufo

Unsure if @casey is still maintainer. If not, we commit to maintain this
fork so long as MFEK needs it.

The licensing change is because all work on MFEK is Apache-2 licensed,
not CC-1.0.